requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
sqlmodel>=0.0.24
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1

# Dev & Tests
pytest>=8.2
httpx>=0.27
ruff>=0.5
mypy>=1.11
